kvserver: forget Raft leader on removal in mixed-version
In 23.1, if the leader replica is removed from the range, the first
replica in the descriptor would campaign using PreVote.

In 23.2, we enabled CheckQuorum, where replicas won't grant prevotes if
they've heard from the (removed) leader recently. We also changed the
logic to call an election immediately, without PreVote.

However, in mixed-version clusters, a 23.1 campaigner can no longer get
prevotes from 23.2 nodes. This patch makes 23.2 nodes forget the current
leader when they apply the leader removal, which allows them to grant
prevotes in this case. This is not necessary once the upgrade is
finalized, since we no longer use PreVote in this case.

Epic: none
Release note: None